refactor(DeckFactory): 스트림을 이용하여 코드 간결화 및 테스트 코드 추가
