+++ src/test/java/model/deck/DeckTest.java
@@ -1,10 +1,9 @@
-package model;
+package model.deck;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.ArrayList;
 import java.util.List;
-import model.deck.Deck;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
